+++ example/pages/search/index.js
@@ -1,7 +1,18 @@
 Page({
   data: {
     value1: '',
-    value2: '初始文案'
+    value2: '初始文案',
+    value3: '',
+    searchType: '全部',
+    menu: [
+      {
+        content: '全部'
+      }, {
+        content: '订单号'
+      }, {
+        content: '退款单号'
+      }
+    ]
   },
   focus () {
     console.log('聚焦')
@@ -20,5 +31,10 @@ Page({
   },
   change (e) {
     console.log('输入', e)
+  },
+  changeSearchType ({ detail: { item } }) {
+    this.setData({
+      searchType: item.content
+    })
   }
 })